pyproject.toml
[project]
name = "tw-homedog"
version = "0.1.0"
description = "Taiwan Real Estate Smart Listing Notifier"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot[job-queue]>=22.6",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tw_homedog"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
